app/requirements.txt
cloudflare==4.3.1
crispy-bootstrap5==2025.6
Django==5.2.6
django-crispy-forms==2.4
django-debug-toolbar==4.4.6
django-environ==0.12.0
django-extensions==4.1
dramatiq==1.18.0
gunicorn==23.0.0
ipython==9.5.0
redis==6.4.0
requests==2.32.5
